pyproject.toml
[project]
name = "ember-backend"
version = "0.1.0"
description = "Ember Pulse backend service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12,<1.0.0",
    "fido2>=1.1.3,<2.0.0",
    "pydantic-settings>=2.10.1,<3.0.0",
    "pyjwt>=2.10.1,<3.0.0",
    "pyodbc>=5.2.0,<6.0.0",
    "sqlalchemy>=2.0.43,<3.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.scripts]
ember-backend = "ember_backend.main:main"

[dependency-groups]
dev = [
    "httpx>=0.28.1,<1.0.0",
    "pytest>=8.4.1,<9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
